Remove unnecessary abstraction
